refactor(ui): split menuController.ts into render/nav/fill modules
Address fallow split-high-impact target (was 694 LOC, 4 dependents):
- menuRender.ts: pure hex-menu markup (renderMenuHtml + MenuState)
- menuNav.ts: stateless keyboard-navigation DOM helpers
- menuFill.ts: custom-fill inline-input wiring (emit/close callbacks)
- controller core down to 398 LOC; dependents per module now minimal

Fallow fully green: dead code 0, complexity findings 0, dupes 0, targets 0. 954 tests passing.
